+++ src/packages/collapseitem/__test__/collapseitem.spec.js
@@ -1,0 +1,5 @@
+import { shallowMount, mount } from '@vue/test-utils';
+import CollapseItem from '../collapseitem.vue';
+import Vue from 'vue';
+
+describe('CollapseItem.vue', () => {});