+++ src/packages/collapseitem/demo.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="demo-list"></div>
+</template>
+<script>
+export default {
+  data() {
+    return {};
+  },
+  methods: {}
+};
+</script>