+++ src/packages/coupon/index.js
@@ -1,0 +1,8 @@
+import Coupon from './coupon.vue';
+import './coupon.scss';
+
+Coupon.install = function(Vue) {
+  Vue.component(Coupon.name, Coupon);
+};
+
+export default Coupon;