+++ src/packages/picker/picker.vue
@@ -9,7 +9,7 @@
       <div class="nut-picker-panel">
         <template v-for="(item, index) of listData">
           <nut-picker-slot
-            :ref="`picer-slot-${index}`"
+            :ref="`picker-slot-${index}`"
             :default-value="chooseValueData[index]"
             :is-update="isUpdate"
             :list-data="item"
@@ -68,12 +68,15 @@ export default {
       this.chooseValueData = [...this.defaultValueData];
       this.cacheValueData = [...this.defaultValueData];
       this.$emit('confirm', this.cacheValueData);
+    },
+    listData: function() {
+      this.init();
     }
   },
   methods: {
     updateChooseValue(self, index, value) {
       self.cacheValueData.splice(index, 1, value);
-      let ref = `picer-slot-${index}`;
+      let ref = `picker-slot-${index}`;
       self.$refs[ref] && self.$refs[ref][0].updateTransform(value);
     },
 
@@ -89,24 +92,27 @@ export default {
       this.chooseValueData = [...this.cacheValueData];
       this.$emit('close');
     },
-
     chooseItem(value, index) {
       if (this.cacheValueData[index] !== value) {
         this.cacheValueData[index] = value;
         this.$emit('choose', this, index, value, this.cacheValueData);
       }
+    },
+    init() {
+      if (this.defaultValueData && this.defaultValueData.length) {
+        this.chooseValueData = [...this.defaultValueData];
+      } else {
+        let defaultValueData = [];
+        this.listData.map((item, index) => {
+          defaultValueData.push(item[0]);
+        });
+        this.chooseValueData = [...defaultValueData];
+      }
     }
   },
+
   created() {
-    if (this.defaultValueData && this.defaultValueData.length) {
-      this.chooseValueData = [...this.defaultValueData];
-    } else {
-      let defaultValueData = [];
-      this.listData.map((item, index) => {
-        defaultValueData.push(item[0]);
-      });
-      this.chooseValueData = [...defaultValueData];
-    }
+    this.init();
   }
 };
 </script>