+++ src/packages/tab/tab.vue
@@ -83,10 +83,19 @@ export default {
       activeIndex: this.defIndex,
       initX: '0',
       navWidth: 0,
-      tapWidth: 0
+      tapWidth: 0,
+      smoothFlag: true
     };
   },
   watch: {
+    $route: {
+      handler(val, oldval) {
+        this.smoothFlag = false;
+        this.scrollTab(this.activeIndex);
+      },
+      // 深度观察监听
+      deep: true
+    },
     isScroll() {
       this.updeteTab();
     },
@@ -109,10 +118,11 @@ export default {
       if (this.positionNav === 'left' || this.positionNav === 'right') return true;
     },
     titleClass: function() {
+      const smooth = this.smoothFlag ? 'nut-tab-title-smooth' : '';
       if (this.positionNav == 'top') {
-        return 'nut-tab-title';
+        return 'nut-tab-title' + ' ' + smooth;
       }
-      return 'nut-tab-title-' + this.positionNav + 'nav';
+      return 'nut-tab-title-' + this.positionNav + 'nav' + ' ' + smooth;
     },
     navBarClass: function() {
       if (this.positionNav == 'top') {
@@ -183,13 +193,13 @@ export default {
             badge: attrs['badge'] || false
           };
           this.tabTitleList.push(item);
-          let slotElm = slot[i].elm;
-          if (slotElm) {
-            slotElm.classList.add('hide');
-            if (this.activeIndex == i) {
-              slotElm.classList.remove('hide');
-            }
-          }
+          //   let slotElm = slot[i].elm;
+          //   if (slotElm) {
+          //     slotElm.classList.add('hide');
+          //     if (this.activeIndex == i) {
+          //       slotElm.classList.remove('hide');
+          //     }
+          //   }
         }
       }
       this.$nextTick(() => {
@@ -209,6 +219,7 @@ export default {
           this.initX = parseInt(this.navWidth * this.defIndex);
           this.tapWidth = tapWidth / 2 - this.navWidth / 2;
         }
+        this.scrollTab(this.activeIndex);
       });
     },
     findParent(event, myclass) {
@@ -225,29 +236,32 @@ export default {
     },
     switchTab: function(index, event, disable) {
       if (!disable) {
+        this.smoothFlag = true;
         this.activeIndex = index;
         // this.initX = parseInt(this.navWidth * index);
-        if (this.lineWidth > 0 && this.lineWidth < this.navWidth) {
-          this.initX = parseInt((this.navWidth - this.lineWidth) / 2 + this.navWidth * index);
-        } else {
-          this.initX = parseInt(this.navWidth * index);
-        }
-        if (this.positionNav == 'top' || this.positionNav == 'bottom') {
-          this.$refs.navlist.scroll(this.initX - this.tapWidth, 0);
-        } else {
-          this.$refs.navlist.scroll(0, this.initX - this.tapWidth);
-        }
-        let nutTab = this.findParent(event, 'nut-tab-part');
-        let items = this.$refs.items.children;
-        for (let i = 0; i < items.length; i++) {
-          if (i == index) {
-            items[i].classList.remove('hide');
-          } else {
-            items[i].classList.add('hide');
-          }
-        }
+        this.scrollTab(index);
         this.$emit('tab-switch', index, event);
         this.$emit('tabSwitch', index, event); //兼容以前驼峰法命名
+      }
+    },
+    scrollTab(index) {
+      if (this.lineWidth > 0 && this.lineWidth < this.navWidth) {
+        this.initX = parseInt((this.navWidth - this.lineWidth) / 2 + this.navWidth * index);
+      } else {
+        this.initX = parseInt(this.navWidth * index);
+      }
+      if (this.positionNav == 'top' || this.positionNav == 'bottom') {
+        this.$refs.navlist.scroll(this.initX - this.tapWidth, 0, 0);
+      } else {
+        this.$refs.navlist.scroll(0, this.initX - this.tapWidth);
+      }
+      let items = this.$refs.items.children;
+      for (let i = 0; i < items.length; i++) {
+        if (i == index) {
+          items[i].classList.remove('hide');
+        } else {
+          items[i].classList.add('hide');
+        }
       }
     }
   }